requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
click==7.1.2              # via -r requirements.txt, pip-tools
coverage==5.0             # via -r requirements.txt
mock==3.0.5               # via -r requirements.txt
mysql-connector-python==8.0.18  # via -r requirements.txt
pip-tools==5.2.1          # via -r requirements.txt
protobuf==3.11.1          # via -r requirements.txt, mysql-connector-python
six==1.13.0               # via -r requirements.txt, mock, pip-tools, protobuf

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
